[core] Fixed crashes for composite components
- Fixed crashes that happen when returning null from composite components
- Fixed crashes that happen when replacing composite components
- Added tests for composite component refs, replacements and removal
